separate assessment in redux store into the regular assessment and gradingAssessment
